requirements.txt
asgiref==3.10.0
boto3==1.40.50
botocore==1.40.50
crispy-bootstrap4==2025.6
Django==5.2.7
django-crispy-forms==2.4
django-storages==1.14.6
gunicorn==23.0.0
jmespath==1.0.1
packaging==25.0
pillow==11.3.0
python-dateutil==2.9.0.post0
s3transfer==0.14.0
six==1.17.0
sqlparse==0.5.3
urllib3==2.5.0
